[apps/desktop] move state handling to a utility
