+++ apps/web/components/Pages/Feed/Content/utilities/sendMessageWrapper.tsx
@@ -7,7 +7,7 @@ import {
   UploadedFile,
 } from '@linen/types';
 import { username } from 'serializers/user';
-import { FeedResponse } from '../types';
+import { FeedResponse } from '../../types';
 import { v4 as uuid } from 'uuid';
 import debounce from '@linen/utilities/debounce';
 import { StartSignUpFn } from 'contexts/Join';
